pyproject.toml
[tool.poetry]
name = "goit-pyweb-hw6"
version = "0.1.0"
description = ""
authors = ["Ilya <159926377+TechnoGoku@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
